fix(worker): computeRetryDelay 去掉冗余 uint(n) 转换以静默 gosec G115
gosec v2.22.0 的 G115 (CWE-190 integer overflow) 规则做模式匹配，
看到 int->uint 转换就报告，不做 dataflow 分析，看不到上游已有
"n<0 -> n=0" + "n>=30 -> return max" 两道 guard 把 n 约束在 [0,29]。

Go 1.13 起 shift count 允许有符号整数；既然上游已保证 n>=0，转
uint 就是历史遗留写法，直接用 1<<n 即可：

- 行为完全等价（n 已被约束到 [0,29]，无溢出风险）
- gosec EXIT=0，security workflow 第一次定时跑误报消除
- 注释段 (60 行)"溢出 1<<n 变负数"现在和代码字面量一致

背景：security.yml 走 schedule (cron)，从 workflow 加入到 2026-05-25
首次 fire 期间从未在 PR 上跑过；这条告警不是 dependabot bump 引入，
而是 gosec G115 规则历史首次扫到这段老代码。

+++ internal/worker/server.go
@@ -70,7 +70,7 @@ func computeRetryDelay(n int, base, max time.Duration) time.Duration {
 	if n >= 30 {
 		return max
 	}
-	delay := time.Duration(1<<uint(n)) * base
+	delay := time.Duration(1<<n) * base
 	if delay <= 0 || delay > max {
 		return max
 	}
